requirements.txt
beautifulsoup4
newspaper4k
feedgen
lxml
requests
google-generativeai
adblockparser
python-dotenv
